+++ miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/dao/GZOrderDao.java
@@ -10,7 +10,7 @@ public interface GZOrderDao {
      * @param id
      * @return
      */
-    List<TGzOrder> findMyOrderList(Integer id);
+    List<TGzOrder> findMyOrderList(Long id);
 
     List<TGzOrder> selectByUserIdAndSubjectIdAndStatusCreateTimeDesc(Long userId, Long subjectId, int i);
